Refactor CSS styles and add orthoGraphyModal class
This commit refactors the CSS styles in main.css and adds the orthoGraphyModal class. The orthoGraphyModal class is positioned relative with a z-index of 10000. Additionally, the color of the modal container body and headings inside the orthoGraphyModal class is set to black.

Added .gitignore file

This commit adds a .gitignore file to the repository, which ignores the .DS_Store file.